work in progress: remove dependency shaderFloat64

+++ shader-src/fontShader.vert
@@ -1,6 +1,6 @@
 #version 450
 
-layout(location = 0) in dvec2 inPosition;
+layout(location = 0) in vec2 inPosition;
 layout(location = 1) in float inTexX;
 layout(location = 2) in float inTexWidth;
 layout(location = 3) in float inSize;
